Add check to not delete keys accidentally
If an engine is being used, the engine will not create a copy of the
used key in most cases and thus, deletion would delete used keys instead
of their copies. This fixes it

+++ agent/lib/perl/CertNanny/Keystore.pm
@@ -1614,9 +1614,11 @@ sub sendrequest {
 	my $enroller = $self->get_enroller();
 	$enroller->enroll(%options);
 
-    unlink $requestkeyfile;
-    unlink $oldkeyfile if (defined $oldkeyfile);
-    unlink $oldcertfile if (defined $oldcertfile);
+    unless($self->hasEngine()) {
+        unlink $requestkeyfile;
+        unlink $oldkeyfile if (defined $oldkeyfile);
+        unlink $oldcertfile if (defined $oldcertfile);
+    }
 
     if (-r $newcertfile) {
 	# successful installation of the new certificate.
